+++ Gambit/Board.h
@@ -16,12 +16,12 @@ public:
     Enums::Piece boardType;
     Enums::Color boardColor;
 
-    unsigned long board;
-    unsigned long bitMask;
+    unsigned long long board;
+    unsigned long long bitMask;
     std::bitset<64> bitBoard;
 
     Board();
-    Board(Enums::Color boardColor,Enums::Piece boardType, unsigned long boardValue);
+    Board(Enums::Color boardColor,Enums::Piece boardType, unsigned long long boardValue);
     bool setBitBoard();
     bool setBitMask();
     bool placePiece(std::string rank, std::string file);
